requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
easyocr==1.7.2
filelock==3.20.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2image==1.17.0
Pillow==10.1.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
PyPDF2==3.0.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.10.0
PyYAML==6.0.3
reportlab==4.0.7
requests==2.31.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
SQLAlchemy==2.0.45
sympy==1.14.0
tifffile==2025.12.12
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
Werkzeug==3.0.1
wsproto==1.3.2
